tree: move `diff_summary()` to `MergedTree`
