+++ lib/tests/test_diff_summary.rs
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 use jj_lib::matchers::{EverythingMatcher, FilesMatcher};
+use jj_lib::merged_tree::{DiffSummary, MergedTree};
 use jj_lib::repo_path::RepoPath;
-use jj_lib::tree::DiffSummary;
 use test_case::test_case;
 use testutils::TestRepo;
 
@@ -48,7 +48,7 @@ fn test_types(use_git: bool) {
     );
 
     assert_eq!(
-        tree1.diff_summary(&tree2, &EverythingMatcher),
+        MergedTree::Legacy(tree1).diff_summary(&MergedTree::Legacy(tree2), &EverythingMatcher),
         DiffSummary {
             modified: vec![modified_path],
             added: vec![added_path],
@@ -70,7 +70,8 @@ fn test_tree_file_transition(use_git: bool) {
     let tree2 = testutils::create_tree(repo, &[(&dir_path, "contents")]);
 
     assert_eq!(
-        tree1.diff_summary(&tree2, &EverythingMatcher),
+        MergedTree::Legacy(tree1.clone())
+            .diff_summary(&MergedTree::Legacy(tree2.clone()), &EverythingMatcher),
         DiffSummary {
             modified: vec![],
             added: vec![dir_path.clone()],
@@ -78,7 +79,7 @@ fn test_tree_file_transition(use_git: bool) {
         }
     );
     assert_eq!(
-        tree2.diff_summary(&tree1, &EverythingMatcher),
+        MergedTree::Legacy(tree2).diff_summary(&MergedTree::Legacy(tree1), &EverythingMatcher),
         DiffSummary {
             modified: vec![],
             added: vec![dir_file_path],
@@ -128,7 +129,8 @@ fn test_sorting(use_git: bool) {
     );
 
     assert_eq!(
-        tree1.diff_summary(&tree2, &EverythingMatcher),
+        MergedTree::Legacy(tree1.clone())
+            .diff_summary(&MergedTree::Legacy(tree2.clone()), &EverythingMatcher),
         DiffSummary {
             modified: vec![a_path.clone(), f_a_path.clone(), f_f_a_path.clone()],
             added: vec![
@@ -143,7 +145,7 @@ fn test_sorting(use_git: bool) {
         }
     );
     assert_eq!(
-        tree2.diff_summary(&tree1, &EverythingMatcher),
+        MergedTree::Legacy(tree2).diff_summary(&MergedTree::Legacy(tree1), &EverythingMatcher),
         DiffSummary {
             modified: vec![a_path, f_a_path, f_f_a_path],
             added: vec![],
@@ -166,7 +168,8 @@ fn test_matcher_dir_file_transition(use_git: bool) {
 
     let matcher = FilesMatcher::new(&[a_path.clone()]);
     assert_eq!(
-        tree1.diff_summary(&tree2, &matcher),
+        MergedTree::Legacy(tree1.clone())
+            .diff_summary(&MergedTree::Legacy(tree2.clone()), &matcher),
         DiffSummary {
             modified: vec![],
             added: vec![],
@@ -174,7 +177,8 @@ fn test_matcher_dir_file_transition(use_git: bool) {
         }
     );
     assert_eq!(
-        tree2.diff_summary(&tree1, &matcher),
+        MergedTree::Legacy(tree2.clone())
+            .diff_summary(&MergedTree::Legacy(tree1.clone()), &matcher),
         DiffSummary {
             modified: vec![],
             added: vec![a_path.clone()],
@@ -184,7 +188,8 @@ fn test_matcher_dir_file_transition(use_git: bool) {
 
     let matcher = FilesMatcher::new(&[a_a_path.clone()]);
     assert_eq!(
-        tree1.diff_summary(&tree2, &matcher),
+        MergedTree::Legacy(tree1.clone())
+            .diff_summary(&MergedTree::Legacy(tree2.clone()), &matcher),
         DiffSummary {
             modified: vec![],
             added: vec![a_a_path.clone()],
@@ -192,7 +197,8 @@ fn test_matcher_dir_file_transition(use_git: bool) {
         }
     );
     assert_eq!(
-        tree2.diff_summary(&tree1, &matcher),
+        MergedTree::Legacy(tree2.clone())
+            .diff_summary(&MergedTree::Legacy(tree1.clone()), &matcher),
         DiffSummary {
             modified: vec![],
             added: vec![],
@@ -202,7 +208,8 @@ fn test_matcher_dir_file_transition(use_git: bool) {
 
     let matcher = FilesMatcher::new(&[a_path.clone(), a_a_path.clone()]);
     assert_eq!(
-        tree1.diff_summary(&tree2, &matcher),
+        MergedTree::Legacy(tree1.clone())
+            .diff_summary(&MergedTree::Legacy(tree2.clone()), &matcher),
         DiffSummary {
             modified: vec![],
             added: vec![a_a_path.clone()],
@@ -210,7 +217,7 @@ fn test_matcher_dir_file_transition(use_git: bool) {
         }
     );
     assert_eq!(
-        tree2.diff_summary(&tree1, &matcher),
+        MergedTree::Legacy(tree2).diff_summary(&MergedTree::Legacy(tree1), &matcher),
         DiffSummary {
             modified: vec![],
             added: vec![a_path],
@@ -247,7 +254,8 @@ fn test_matcher_normal_cases(use_git: bool) {
 
     let matcher = FilesMatcher::new(&[a_path.clone(), z_path.clone()]);
     assert_eq!(
-        tree1.diff_summary(&tree2, &matcher),
+        MergedTree::Legacy(tree1.clone())
+            .diff_summary(&MergedTree::Legacy(tree2.clone()), &matcher),
         DiffSummary {
             modified: vec![a_path.clone()],
             added: vec![z_path.clone()],
@@ -255,7 +263,8 @@ fn test_matcher_normal_cases(use_git: bool) {
         }
     );
     assert_eq!(
-        tree2.diff_summary(&tree1, &matcher),
+        MergedTree::Legacy(tree2.clone())
+            .diff_summary(&MergedTree::Legacy(tree1.clone()), &matcher),
         DiffSummary {
             modified: vec![a_path],
             added: vec![],
@@ -265,7 +274,8 @@ fn test_matcher_normal_cases(use_git: bool) {
 
     let matcher = FilesMatcher::new(&[dir1_a_path.clone(), dir2_b_path.clone()]);
     assert_eq!(
-        tree1.diff_summary(&tree2, &matcher),
+        MergedTree::Legacy(tree1.clone())
+            .diff_summary(&MergedTree::Legacy(tree2.clone()), &matcher),
         DiffSummary {
             modified: vec![dir1_a_path.clone()],
             added: vec![dir2_b_path.clone()],
@@ -273,7 +283,7 @@ fn test_matcher_normal_cases(use_git: bool) {
         }
     );
     assert_eq!(
-        tree2.diff_summary(&tree1, &matcher),
+        MergedTree::Legacy(tree2).diff_summary(&MergedTree::Legacy(tree1), &matcher),
         DiffSummary {
             modified: vec![dir1_a_path],
             added: vec![],
